+++ cardano-crypto-leios/CHANGELOG.md
@@ -2,5 +2,4 @@
 
 ## 0.1.0.0
 
-* Initial version: `EbHash` and `LeiosCert` extracted from
-  `ouroboros-consensus` and `cardano-ledger-core` per CIP-0164.
+* Initial version of `Cardano.Crypto.Leios` that introduces `LeiosCert`, `LeiosCommittee`, and `LeiosVoterId` types, as well as main functions to interact with the types: `resolveLeiosVoter`, `getLeiosVoterId`, `aggregateLeiosCert`, and `verifyLeiosCert`  being notable functions.